+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - Increase code font size by 50%.
+
 ## v0.13.0
 
   - Change zoom to 50% on slide preview (presenter display).